GEODE-6195: Check if returned value is caused by a retried putIfAbsent operation
  * Treats a retried putIfAbsent operation a successful one if returned value is the value to be put
  * Move the check value methods to ValueComparisonHelper class.
