+++ geode-core/src/test/java/org/apache/geode/internal/cache/LocalRegionTest.java
@@ -27,6 +27,7 @@ import org.junit.Test;
 import org.apache.geode.CancelCriterion;
 import org.apache.geode.cache.EntryNotFoundException;
 import org.apache.geode.cache.Operation;
+import org.apache.geode.cache.RegionAttributes;
 import org.apache.geode.cache.client.internal.ServerRegionProxy;
 
 public class LocalRegionTest {
@@ -35,6 +36,7 @@ public class LocalRegionTest {
   private ServerRegionProxy serverRegionProxy;
   private Operation operation;
   private CancelCriterion cancelCriterion;
+  private RegionAttributes regionAttributes;
 
   private final Object key = new Object();
   private final String value = "value";
@@ -45,12 +47,14 @@ public class LocalRegionTest {
     event = mock(EntryEventImpl.class);
     serverRegionProxy = mock(ServerRegionProxy.class);
     cancelCriterion = mock(CancelCriterion.class);
+    regionAttributes = mock(RegionAttributes.class);
 
     when(region.getServerProxy()).thenReturn(serverRegionProxy);
     when(event.isFromServer()).thenReturn(false);
     when(event.getKey()).thenReturn(key);
     when(event.getRawNewValue()).thenReturn(value);
     when(region.getCancelCriterion()).thenReturn(cancelCriterion);
+    when(region.getAttributes()).thenReturn(regionAttributes);
   }
 
   @Test
@@ -91,7 +95,7 @@ public class LocalRegionTest {
   public void checkPutIfAbsentResultThrowsIfEventHasRetriedButResultNotHaveSameValue() {
     Object result = new Object();
     when(event.hasRetried()).thenReturn(true);
-    when(region.putIfAbsentResultHasSameValue(value, result)).thenReturn(false);
+    when(region.putIfAbsentResultHasSameValue(true, value, result)).thenReturn(false);
     doCallRealMethod().when(region).checkPutIfAbsentResult(event, value, result);
 
     region.checkPutIfAbsentResult(event, value, result);
@@ -101,43 +105,139 @@ public class LocalRegionTest {
   public void checkPutIfAbsentResultSucceedsIfEventHasRetriedAndResultHasSameValue() {
     Object result = new Object();
     when(event.hasRetried()).thenReturn(true);
-    when(region.putIfAbsentResultHasSameValue(value, result)).thenReturn(true);
+    when(region.putIfAbsentResultHasSameValue(true, value, result)).thenReturn(true);
     doCallRealMethod().when(region).checkPutIfAbsentResult(event, value, result);
 
     region.checkPutIfAbsentResult(event, value, result);
 
     verify(event).hasRetried();
-    verify(region).putIfAbsentResultHasSameValue(value, result);
+    verify(region).putIfAbsentResultHasSameValue(true, value, result);
   }
 
   @Test
   public void putIfAbsentResultHasSameValueReturnTrueIfResultIsInvalidTokenAndValueToBePutIsNull() {
-    when(region.putIfAbsentResultHasSameValue(null, Token.INVALID)).thenCallRealMethod();
+    when(region.putIfAbsentResultHasSameValue(true, null, Token.INVALID)).thenCallRealMethod();
 
-    assertThat(region.putIfAbsentResultHasSameValue(null, Token.INVALID)).isTrue();
+    assertThat(region.putIfAbsentResultHasSameValue(true, null, Token.INVALID)).isTrue();
   }
 
   @Test
   public void putIfAbsentResultHasSameValueReturnFalseIfResultIsInvalidTokenAndValueToBePutIsNotNull() {
-    when(region.putIfAbsentResultHasSameValue(value, Token.INVALID)).thenCallRealMethod();
+    when(region.putIfAbsentResultHasSameValue(true, value, Token.INVALID)).thenCallRealMethod();
 
-    assertThat(region.putIfAbsentResultHasSameValue(value, Token.INVALID)).isFalse();
+    assertThat(region.putIfAbsentResultHasSameValue(true, value, Token.INVALID)).isFalse();
   }
 
   @Test
   public void putIfAbsentResultHasSameValueReturnTrueIfResultHasSameValue() {
     Object result = "value";
-    when(region.putIfAbsentResultHasSameValue(value, result)).thenCallRealMethod();
+    when(region.putIfAbsentResultHasSameValue(true, value, result)).thenCallRealMethod();
 
-    assertThat(region.putIfAbsentResultHasSameValue(value, result)).isTrue();
+    assertThat(region.putIfAbsentResultHasSameValue(true, value, result)).isTrue();
+    verify(region, never()).getAttributes();
   }
 
   @Test
   public void putIfAbsentResultHasSameValueReturnFalseIfResultDoesNotHaveSameValue() {
     Object result = "differentValue";
-    when(region.putIfAbsentResultHasSameValue(value, result)).thenCallRealMethod();
+    when(region.putIfAbsentResultHasSameValue(true, value, result)).thenCallRealMethod();
 
-    assertThat(region.putIfAbsentResultHasSameValue(value, result)).isFalse();
+    assertThat(region.putIfAbsentResultHasSameValue(true, value, result)).isFalse();
+    verify(region, never()).getAttributes();
+  }
+
+  @Test
+  public void putIfAbsentResultHasSameValueChecksRegionAttributesIfNotFromClient() {
+    Object result = "value";
+    when(region.putIfAbsentResultHasSameValue(false, value, result)).thenCallRealMethod();
+
+    assertThat(region.putIfAbsentResultHasSameValue(false, value, result)).isTrue();
+    verify(region).getAttributes();
+  }
+
+  @Test
+  public void putIfAbsentResultHasSameValueReturnFalseIfResultDoesNotHaveSameValueAndNotFromClient() {
+    Object oldValue = "differentValue";
+    Object result = new VMCachedDeserializable(EntryEventImpl.serialize(oldValue));
+    when(region.putIfAbsentResultHasSameValue(false, value, result)).thenCallRealMethod();
+
+    assertThat(region.putIfAbsentResultHasSameValue(false, value, result)).isFalse();
+    verify(region).getAttributes();
+  }
+
+  @Test
+  public void putIfAbsentResultHasSameValueReturnTrueIfResultHasSameValueAndNotFromClient() {
+    Object result = new VMCachedDeserializable(EntryEventImpl.serialize(value));
+    when(region.putIfAbsentResultHasSameValue(false, value, result)).thenCallRealMethod();
+
+    assertThat(region.putIfAbsentResultHasSameValue(false, value, result)).isTrue();
+    verify(region).getAttributes();
+  }
+
+  @Test
+  public void bridgePutIfAbsentResultHasSameValueCanCheckValueForObject() {
+    Object result = "value";
+    byte[] valueToBePut = EntryEventImpl.serialize(value);
+    when(region.bridgePutIfAbsentResultHasSameValue(valueToBePut, true, result))
+        .thenCallRealMethod();
+
+    assertThat(region.bridgePutIfAbsentResultHasSameValue(valueToBePut, true, result)).isTrue();
+    verify(region).getAttributes();
+  }
+
+  @Test
+  public void bridgePutIfAbsentResultHasSameValueCanCheckValueForNonObjectByteArray() {
+    byte[] valueToBePut = {0, 1, 2, 3};
+    Object result = new VMCachedDeserializable(EntryEventImpl.serialize(valueToBePut));
+    when(region.bridgePutIfAbsentResultHasSameValue(valueToBePut, false, result))
+        .thenCallRealMethod();
+
+    assertThat(region.bridgePutIfAbsentResultHasSameValue(valueToBePut, false, result)).isTrue();
+    verify(region).getAttributes();
+  }
+
+  @Test
+  public void bridgePutIfAbsentResultHasSameValueCanCheckValueForIntArray() {
+    int[] newValue = {0, 1, 2, 3};
+    byte[] valueToBePut = EntryEventImpl.serialize(newValue);
+    Object result = newValue;
+    when(region.bridgePutIfAbsentResultHasSameValue(valueToBePut, true, result))
+        .thenCallRealMethod();
+
+    assertThat(region.bridgePutIfAbsentResultHasSameValue(valueToBePut, true, result)).isTrue();
+    verify(region).getAttributes();
+  }
+
+  @Test
+  public void bridgePutIfAbsentResultHasSameValueCanCheckValueForArrayOfArray() {
+    String[] array1 = {"0", "1", "2"};
+    String[] array2 = {"3", "4", "5"};
+    String[] array3 = {"7"};
+    String[][] newValue = {array1, array2, array3};
+    byte[] valueToBePut = EntryEventImpl.serialize(newValue);
+    Object result = new VMCachedDeserializable(EntryEventImpl.serialize(newValue));
+    when(region.bridgePutIfAbsentResultHasSameValue(valueToBePut, true, result))
+        .thenCallRealMethod();
+
+    assertThat(region.bridgePutIfAbsentResultHasSameValue(valueToBePut, true, result)).isTrue();
+    verify(region).getAttributes();
+  }
+
+  @Test
+  public void bridgePutIfAbsentResultHasSameValueCanCheckDifferentValuesForArrayOfArray() {
+    String[] array1 = {"0", "1", "2"};
+    String[] array2 = {"3", "4", "5"};
+    String[] array3 = {"7"};
+    String[] array4 = {"8"};
+    String[][] newValue = {array1, array2, array3};
+    String[][] returnedValue = {array1, array2, array4};
+    byte[] valueToBePut = EntryEventImpl.serialize(newValue);
+    Object result = new VMCachedDeserializable(EntryEventImpl.serialize(returnedValue));
+    when(region.bridgePutIfAbsentResultHasSameValue(valueToBePut, true, result))
+        .thenCallRealMethod();
+
+    assertThat(region.bridgePutIfAbsentResultHasSameValue(valueToBePut, true, result)).isFalse();
+    verify(region).getAttributes();
   }
 
 }
